+++ browser_test/jqutil_test.js
@@ -34,46 +34,6 @@ describe("jqutil", function () {
                          ".btw\\:cit._real .tei\\:q._real");
         });
     });
-
-    describe("selectorToElements", function () {
-        it("raises an error on brackets",
-                 function () {
-            assert.Throw(jqutil.selectorToElements.bind(undefined, "abcde[f]"),
-                         Error, "selector is too complex");
-        });
-
-        it("raises an error on parens",
-                 function () {
-            assert.Throw(jqutil.selectorToElements.bind(undefined,
-                                                        "abcde:not(f)"),
-                         Error, "selector is too complex");
-        });
-
-        it("raises an error on classes",
-                 function () {
-            assert.Throw(jqutil.selectorToElements.bind(undefined,
-                                                        "blah.q"),
-                         Error, "selector is too complex for "+
-                         "selectorToElements");
-        });
-
-        it("raises an error on spaces",
-                 function () {
-            assert.Throw(jqutil.selectorToElements.bind(undefined,
-                                                        "p term"),
-                         Error, "selector is not a sequence of > " +
-                         "selectors");
-        });
-
-        it("converts a > sequence",
-                 function () {
-            assert.equal(
-                jqutil.selectorToElements("p > term > foreign")[0].outerHTML,
-                '<div class="p _real"><div class="term _real">'+
-                    '<div class="foreign _real"></div></div></div>');
-        });
-    });
-
 });
 
 });